vertical shift added

+++ src/main/java/metrics/Metrics.java
@@ -4,7 +4,7 @@ public enum Metrics implements Difference{
     Euclidean((p1, p2) -> {
         double res = 0;
         for (int i = 0; i < p1.length; i++) {
-            int d = p1[i] - p2[i];
+            double d = p1[i] - p2[i];
             res += d * d;
         }
         return Math.sqrt(res);
